Hotfixing get weather info #23
